# specs/001-movie-review-website/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Movie Review Website API
  version: 1.0.0
  description: API for browsing movies, posting reviews, and admin management
servers:
  - url: https://api.movie-review-website.com/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect
            profile: User profile
            email: User email
  schemas:
    Movie:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        year:
          type: integer
        actors:
          type: array
          items:
            type: string
        director:
          type: string
        image:
          type: string
          description: Base64 encoded image
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - title
        - year
        - director
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
paths:
  /movies:
    get:
      summary: Browse movies
      description: Retrieve a list of movies, optionally filtered by title or year
      parameters:
        - name: title
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of movies
          content:
            application/json:
              schema:
                type: object
                properties:
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Movie'
                  total:
                    type: integer
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new movie
      description: Admin only - Add a new movie to the database
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movie'
      responses:
        '201':
          description: Movie created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
        '400':
          description: Bad request
...